requirements.txt
tqdm==4.32.2
tensorflow==1.13.2
six
deeppavlov
python-Levenshtein
matplotlib
pandas==0.24.2
pybind11==2.2.3
numpy==1.16.4
xlwt
python-Levenshtein
gensim
gevent
tensor2tensor
aiohttp
tldextract
requests_html